ds1 editor: moved stat management methods into state.go
